refactor: extract sqlite/mod.rs into 5 focused modules (PR-2c)
